add css variables and fix rules
