+++ package.json
@@ -36,6 +36,7 @@
     "webpack-merge": "5.8.0"
   },
   "dependencies": {
+    "normalize": "0.3.1",
     "swiper": "10.3.0"
   }
 }
