+++ src/components/App.scss
@@ -1,3 +1,5 @@
+@import '../vendor/variables';
+
 .container {
   position: relative;
   box-sizing: border-box;
@@ -5,6 +7,10 @@
   width: 100%;
   margin: auto;
   padding: 0 80px;
+
+  @media all and (max-width: 320px) {
+    padding: 0 20px;
+  }
 }
 
 .swiper {
@@ -15,7 +21,7 @@
 .swiper-slide {
   text-align: center;
   font-size: 18px;
-  background-color: transparent;
+  background-color: $transparent;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -23,16 +29,17 @@
 
 .outer-swiper {
   min-height: 300px;
-  // border: 1px solid red;
-  background-color: transparent;
+  background-color: $transparent;
   margin-left: 80px;
+
+  @media all and (max-width: 320px) {
+    min-height: 100px;
+    margin-left: 0;
+  }
 }
 
 .swiper-slide {
-  // min-height: 200px;
-  // border: 1px solid rgb(60, 255, 0);
-  // background-color: #371355;
-  background-color: transparent;
+  background-color: $transparent;
 }
 
 .info {
@@ -40,92 +47,76 @@
   gap: 397px;
   align-items: center;
   padding: 165px 0 90px 0;
-}
 
-.info__title {
-  color: var(--Black-blue, #42567a);
-  font-family: 'PT Sans', sans-serif;
-  font-size: 56px;
-  font-weight: 700;
-  line-height: 120%;
-  margin: 0;
-  width: 410px;
-  padding: 0 0 0 73px;
-  border-left: 7px solid transparent;
-  border-image: linear-gradient(#3877ee, #ef5da8);
-  -moz-border-image: -moz-linear-gradient(#3877ee, #ef5da8);
-  -webkit-border-image: -webkit-linear-gradient(#3877ee, #ef5da8);
-  border-image-slice: 1;
-}
+  &__title {
+    color: $black-blue;
+    font-size: 56px;
+    font-weight: 700;
+    line-height: 120%;
+    margin: 0;
+    width: 410px;
+    padding: 0 0 0 73px;
+    border-left: 7px solid transparent;
+    border-image: linear-gradient($helper-color, $fuschia-100);
+    border-image-slice: 1;
+  }
 
-.info__event-type {
-  margin: 0;
-  color: var(--Black-blue, #42567a);
-  font-family: 'PT Sans', sans-serif;
-  font-size: 20px;
-  font-weight: 700;
-  line-height: 30px;
+  &__event-type {
+    margin: 0;
+    color: $black-blue;
+    font-size: 20px;
+    font-weight: 700;
+    line-height: 30px;
+  }
+
+  @media all and (max-width: 320px) {
+    display: block;
+    padding: 60px 0 56px 0;
+
+    &__event-type {
+      display: none;
+    }
+
+    &__title {
+      border-left: none;
+
+      font-size: 20px;
+      padding: 0 0 0 0;
+    }
+  }
 }
 
 .span {
   text-align: center;
-  font-family: 'PT Sans', sans-serif;
   font-size: 200px;
   font-style: normal;
   font-weight: 700;
   line-height: 160px;
   letter-spacing: -4px;
-}
 
-.span_1 {
-  color: var(--iris-100, #5d5fef);
-  padding-left: 218px;
-}
-
-.span_2 {
-  color: var(--fuschia-100, #ef5da8);
-  padding-left: 93px;
-}
-
-@media all and (max-width: 320px) {
-  .container {
-    padding: 0 20px;
-  }
-
-  .info {
-    display: block;
-    padding: 60px 0 56px 0;
-  }
-
-  .info__event-type {
-    display: none;
-  }
-
-  .info__title {
-    border-left: none;
-
-    font-size: 20px;
-    padding: 0 0 0 0;
-  }
-
-  .span {
+  @media all and (max-width: 320px) {
     display: inline-block;
     font-size: 56px;
     line-height: normal;
     letter-spacing: -1.12px;
     border-bottom: 1px solid #c7cdd9;
   }
+}
 
-  .span_1 {
+.span_1 {
+  color: $iris-100;
+  padding-left: 218px;
+
+  @media all and (max-width: 320px) {
     padding: 0 0 60px 0;
   }
+}
 
-  .span_2 {
+.span_2 {
+  color: $fuschia-100;
+  padding-left: 93px;
+
+  @media all and (max-width: 320px) {
     padding: 0 0 60px 20px;
-  }
-
-  .outer-swiper {
-    min-height: 100px;
-    margin-left: 0;
   }
 }
