+++ src/components/Circle/Circle.scss
@@ -1,49 +1,41 @@
-.circle {
-  // width: 1200px;
-  // height: 800px;
+@import '../../vendor/variables';
 
+.circle {
   width: 100%;
   height: 100%;
   display: flex;
   justify-content: center;
   align-items: center;
-  // position: relative;
   top: 0;
   position: absolute;
-  border: 1px solid var(--Black-blue, rgba(66, 86, 122, 0.2));
+  border: 1px solid $black-blue-dimmed;
   border-bottom: none;
   border-top: none;
+
+  @media all and (max-width: 320px) {
+    display: none;
+  }
 }
 
-// .circle::before,
 .circle::after {
   content: '';
   display: block;
-  // width: 2px;
   height: 110vh;
   position: absolute;
   top: -94px;
   left: 50%;
-  // background-color: red;
-  border: 1px solid var(--Black-blue, rgba(66, 86, 122, 0.2));
+  border: 1px solid $black-blue-dimmed;
 }
-
-// .circle::before {
-//   transform: rotate(90deg);
-//   height: 100%;
-// }
 
 .circle::before {
   content: '';
   display: block;
-  // width: 2px;
   height: 1442px;
   position: absolute;
   top: -241px;
   left: 50%;
-  // background-color: red;
   transform: rotate(90deg);
-  border: 1px solid var(--Black-blue, rgba(66, 86, 122, 0.2));
+  border: 1px solid $black-blue-dimmed;
 }
 
 .wrapper {
@@ -52,7 +44,7 @@
   border-radius: 50%;
   position: relative;
   margin-bottom: 190px;
-  border: 1px solid var(--Black-blue, rgba(66, 86, 122, 0.2));
+  border: 1px solid $black-blue-dimmed;
 }
 
 .point {
@@ -61,13 +53,13 @@
   border-radius: 50%;
   position: absolute;
   transition: 0.3s transform;
-  border: 1px solid #42567a;
-  background-color: #42567a;
-}
+  border: 1px solid $black-blue;
+  background-color: $black-blue;
 
-.point_active {
-  transform: scale(7);
-  background-color: transparent;
+  &_active {
+    transform: scale(7);
+    background-color: $transparent;
+  }
 }
 
 .point:hover {
@@ -102,10 +94,4 @@
 .point_6 {
   top: 50%;
   right: -6px;
-}
-
-@media all and (max-width: 320px) {
-  .circle {
-    display: none;
-  }
 }
