+++ src/components/InnerSwiper/InnerSwiper.scss
@@ -1,11 +1,41 @@
+@import '../../vendor/variables';
+
 .swiper-slide {
   flex-direction: column;
+
+  &__year {
+    color: $accent-blue;
+    font-family: 'Bebas Neue', sans-serif;
+    font-size: 25px;
+    font-weight: 400;
+    line-height: 120%;
+    text-transform: uppercase;
+    align-self: start;
+
+    @media all and (max-width: 320px) {
+      font-size: 16px;
+    }
+  }
+
+  &__description {
+    color: $black-blue;
+    font-size: 20px;
+    font-weight: 400;
+    line-height: 30px;
+    text-align: left;
+    align-self: start;
+
+    @media all and (max-width: 320px) {
+      line-height: 30px;
+      font-size: 14px;
+      line-height: 145%;
+    }
+  }
 }
 
 .swiper-button-next {
   background-image: url('../../assets/images/arrow-left.svg');
   transform: rotate(180deg);
-  // background-image: url('../../assets/images/inner-arrow-right.svg');
   background-repeat: no-repeat;
   background-size: 100% auto;
   background-position: center;
@@ -24,36 +54,4 @@
 
 .swiper-button-prev::after {
   display: none;
-}
-
-.swiper-slide__year {
-  color: #3877ee;
-  font-family: 'Bebas Neue', sans-serif;
-  font-size: 25px;
-  font-weight: 400;
-  line-height: 120%;
-  text-transform: uppercase;
-  align-self: start;
-}
-
-.swiper-slide__description {
-  color: var(--Black-blue, #42567a);
-  font-family: 'PT Sans', sans-serif;
-  font-size: 20px;
-  font-weight: 400;
-  line-height: 30px;
-  text-align: left;
-  align-self: start;
-}
-
-@media all and (max-width: 320px) {
-  .swiper-slide__year {
-    font-size: 16px;
-  }
-
-  .swiper-slide__description {
-    line-height: 30px;
-    font-size: 14px;
-    line-height: 145%;
-  }
 }
